Add LaunchScreen.xib to use as the launch image.
Also removed some duplicate constraints on the signing view and adjusted others.

Note the background color is a solid color at the moment while the actual SignIn view has a striped background because of using a pattern image which may be temporary.

If constraints are updated on the sign in view in the main storyboard it the view will just have to copied and pasted to LaunchScreen.xib