Update methods that are called via selectors to pass the argument as well
Passing the argument makes it easy to identify what will cause this method to be called
For example with "func cancel(sender: UIBarButtonItem)" vs "func cancel()"
Its easy to see that the cancel button is called as a result of pressing a UIBarButtonItem and not some other action in code.
If the sender is not used the method can be updated to be "func cancel(_: UIBarButtonItem)". This still conveys the sender information
Also updated the methods to be "@objc private" to make them private.