Remove statusLabel since it is unused
At the moment updateStatus and showError in ViewController are only used in AuthenticationViewController and SetPasswordViewController

Even if it were used it would not show properly since the label is right above the button and there is currently no code to resize the view based on its contents. Removing for now to simplify things.